requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
scipy==1.15.2
fire==0.7.0
gradio==4.44.1
accelerate==0.29.3
bitsandbytes==0.43.1
diffusers==0.24.0
ftfy==6.2.0
huggingface-hub==0.22.2
Jinja2==3.1.3
numpy==1.26.4
pillow==10.3.0
safetensors==0.4.3
tensorboard==2.16.2
tqdm==4.66.2
transformers==4.40.1
triton==2.2.0
wandb==0.16.6
xformers==0.0.25.post1
torch==2.2.2
torchvision==0.17.2
clip==1.0